+++ server/src/VarDatFile.ts
@@ -2,6 +2,8 @@ import {
 	Range,
 	Hover,
 	HoverParams,
+	FoldingRangeParams,
+	FoldingRange,
 } from 'vscode-languageserver/node';
 
 import {
@@ -32,7 +34,8 @@ export class VarDatFile {
 		
 		const lines = this.workspace.getTextLines( this.filePath );
 		let currentSection = "";
-		let sectionRange = Range.create(0,0,0,0);
+		let headerRange = Range.create(0,0,0,0);
+		let contentsRange = Range.create(0,0,0,0);
 	
 		if( !lines ) {
 			return undefined;
@@ -45,26 +48,28 @@ export class VarDatFile {
 			if( lineText.startsWith("/") ) {
 				const newSectionName = Util.extractSectionNameFromText( lineText );
 				if(newSectionName.length == 0) {
+					headerRange.end.line = i;
 					continue;
 				}
-				else if( currentSection.length > 0 && sectionRange.start.line != sectionRange.end.line ) {
+				else if( currentSection.length > 0 && contentsRange.start.line != contentsRange.end.line ) {
 					// console.log(`new section: ${currentSection} , from ${sectionRange.start.line} to ${sectionRange.end.line}`);
 					this.sectionedDocument.setSectionRange( currentSection, {
-						header: sectionRange,
-						contents: sectionRange
+						header: headerRange,
+						contents: contentsRange
 					} );
 				}
-				sectionRange = Range.create(i+1,0,i+1,0);
+				headerRange = Range.create(i,0,i,0);
+				contentsRange = Range.create(i+1,0,i+1,0);
 				currentSection = newSectionName;
 			}
 			else {
-				sectionRange.end.line = i+1;
+				contentsRange.end.line = i+1;
 			}
 		}
-		if( currentSection.length > 0 && sectionRange.start.line != sectionRange.end.line ) {
+		if( currentSection.length > 0 && contentsRange.start.line != contentsRange.end.line ) {
 			this.sectionedDocument.setSectionRange( currentSection, {
-				header: sectionRange,
-				contents: sectionRange
+				header: headerRange,
+				contents: contentsRange
 			} );
 		}
 		
@@ -112,5 +117,22 @@ export class VarDatFile {
 				`${sectionName} ${offset}`
 			]
 		};
+	}
+
+
+	onFoldingRanges( foldingRangeParam: FoldingRangeParams ) {
+		const foldingRanges: FoldingRange[] = [];
+
+		const sectionedDocument = this.updateSection();
+		if( !sectionedDocument ) {
+			return null;
+		}
+
+		sectionedDocument.sectionMap.forEach((value, key) => {
+			const foldingRange = FoldingRange.create( value.header.start.line, value.contents.end.line-1 );
+			foldingRanges.push(foldingRange);
+		});
+
+		return foldingRanges;
 	}
 }